test(integration): split the boot smoke into three focused files
The single app_smoke_test.dart bundled three distinct boot
assertions into one testWidgets block — main() returning,
routing to the first user-visible screen, and a clean error
budget. Splitting them gives the sharded CI matrix something to
fan out across (3 parallel shards instead of 1), and the file
names make each assertion grep-able when one shard fails.

- app_boot_test.dart: main() finishes and a MaterialApp ends up
  on screen. The narrowest "platform plugins didn't throw"
  canary.
- onboarding_routing_test.dart: with no user data, the first
  screen is OnboardingScreen — guards against the
  hasUserData/routing decision in main.dart silently breaking.
- boot_errors_test.dart: nothing fires FlutterError.onError
  during the boot+settle window. Catches Hive type-id
  collisions, locator dependency cycles, and similar quiet
  regressions.

Each file pays its own simulator boot cost, which is the
trade-off for fan-out. Worth it because each assertion is
genuinely independent and the failure modes don't overlap.
